% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip.R
\name{as.subnet}
\alias{as.subnet}
\title{Convert IPv4 address to its subnet}
\usage{
as.subnet(x, mask, suffix = FALSE)
}
\arguments{
\item{x}{Character vector of IPv4 addresses.}

\item{mask}{Integer between 0 and 32, inclusive. The subnet mask.}

\item{suffix}{Logical. If TRUE, return value should have
"/" + mask concatenated to the end of the string. This produces
Classless Inter-Domain Routing (CIDR) notation for the subnet.}
}
\value{
Character vector.
}
\description{
Convert IPv4 address to its subnet
}
\examples{
as.subnet("192.168.1.1", 16)

as.subnet("192.168.1.1", 24, suffix = TRUE)
}
