% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ban_list}
\alias{ban_list}
\title{IP ban list of suspected spy nodes}
\format{
\subsection{\code{ban_list}}{

A character vector with 423 elements. Most are singleton IP addresses. Six
of them are IP addresses ranges in Classless Inter-Domain Routing (CIDR)
notation.
}
}
\source{
\url{https://github.com/Boog900/monero-ban-list}
}
\usage{
ban_list
}
\description{
A set of suspected spy nodes. See this GitHub issue for more information:
\url{https://github.com/monero-project/meta/issues/1124}
}
\keyword{datasets}
