% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ban_list_v1}
\alias{ban_list_v1}
\alias{ban_list_v2}
\title{IP ban list of suspected spy nodes}
\format{
\subsection{\code{ban_list_v1}}{

A character vector with 423 elements. Most are singleton IP addresses. Six
of them are IP addresses ranges in Classless Inter-Domain Routing (CIDR)
notation.
}

\subsection{\code{ban_list_v2}}{

A character vector with 431 elements.
}

An object of class \code{character} of length 431.
}
\source{
\url{https://github.com/Boog900/monero-ban-list}
}
\usage{
ban_list_v1

ban_list_v2
}
\description{
A set of suspected spy nodes. Ban list version 1 was released in December
2024. Ban list version 2 was released in Janraury 2026. See this GitHub
issue for more information:
\url{https://github.com/monero-project/meta/issues/1124}
}
\keyword{datasets}
