% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{compress.log}
\alias{compress.log}
\title{Extract and compress selected log lines}
\usage{
compress.log(
  bitmonero.dir = c("~/.bitmonero", "C:\\\\ProgramData\\\\bitmonero"),
  output.file = "extracted-xmr-log",
  log.filter = "net.p2p.msg",
  rm.uncompressed.file = TRUE
)
}
\arguments{
\item{bitmonero.dir}{Directory location of the log files}

\item{output.file}{Name of the file that will be created. ".tar.xz" will
be appended to this name}

\item{log.filter}{Regular expression that selects log lines.}

\item{rm.uncompressed.file}{Remove the uncompressed version of the
extracted log file after compression? Does not affect the original log files.}
}
\value{
NULL (invisible)
}
\description{
With default arguments, this function accomplishes the same thing as this
Linux shell call:

\if{html}{\out{<div class="sourceCode">}}\preformatted{grep -I "net.p2p.msg" bitmonero.log* > extracted-xmr-log
for file in bitmonero.log*.tar.gz; do tar xf "$file" --to-command 'egrep -Ha --label="$TAR_FILENAME" "net.p2p.msg" || true' >> extracted-xmr-log; done
xz extracted-xmr-log
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
compress.log()
}
}
