% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{get.p2p.log}
\alias{get.p2p.log}
\title{Parse p2p log files}
\usage{
get.p2p.log(bitmonero.dir = "~/.bitmonero", output.file = NULL)
}
\arguments{
\item{bitmonero.dir}{Directory location of the log files}

\item{output.file}{Optional file to output compressed data}
}
\value{
data.frame of each transaction gossip message for each peer.
}
\description{
Parse p2p log files
}
\examples{
\dontrun{
get.p2p.log()
}
}
