% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip.R
\name{in.ip.set}
\alias{in.ip.set}
\title{Check for IP inclusion in set of IP singletons and ranges}
\usage{
in.ip.set(x, ip.set)
}
\arguments{
\item{x}{Character vector of singleton IP addresses. Can contain
duplicate elements.}

\item{ip.set}{Character vector of IP address singletons and/or ranges.
Ranges should be in Classless Inter-Domain Routing (CIDR) notation, e.g.
"169.254.0.0/24". No duplicate elements allowed. Ranges should not overlap
(see documentation in \code{IP::ip.match}).}
}
\value{
Logical vector , with length same as \code{x}.
}
\description{
Check for IP inclusion in set of IP singletons and ranges
}
\examples{
in.ip.set(c("192.168.5.1", "192.168.4.1", "169.254.0.5"),
  c("169.254.0.0/24", "192.168.1.1", "192.168.5.1"))

}
