% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer-selection.R
\name{peer.selection.collect}
\alias{peer.selection.collect}
\title{Collect peer selection draws}
\usage{
peer.selection.collect(
  fifo.file = "monerod-log.fifo",
  csv.file.suffix = "peer-selection.csv",
  unrestricted.rpc.url = "http://127.0.0.1:18081",
  verbose = 2,
  read.log.wait.time = 0.01,
  white_list.max.size = 1000,
  gray_list.max.size = 5000
)
}
\arguments{
\item{fifo.file}{File name of FIFO ("named pipe") file that \code{monerod}
output is directed to.}

\item{csv.file.suffix}{Suffix for csv files of input data. It should
be the same as used in the \code{peer.selection.collect()} function that
collected the data. The suffix is applied to \verb{white_list-}, \verb{gray_list-},
and \verb{connections-}.}

\item{unrestricted.rpc.url}{URL and port of the \code{monerod} unrestricted RPC.
Default is \verb{http://127.0.0.1:18081}}

\item{verbose}{If 2, print info about every new connection draw. If 1,
print info only about cumulative number of \code{white_list} and \code{gray_list}
draws. If 0, print nothing.}

\item{read.log.wait.time}{Time interval, in seconds, between when the
peer list state is queried through RPC and when the FIFO log file
is read.}

\item{white_list.max.size}{Maximum size of the node's peer \code{white_list}.
This should not be changed under normal circumstances because it is
set in the \code{monerod} code.}

\item{gray_list.max.size}{Maximum size of the node's peer \code{gray_list}.
This should not be changed under normal circumstances because it is
set in the \code{monerod} code.}
}
\value{
NULL (invisible)
}
\description{
TODO
}
\examples{
\dontrun{
peer.selection.collect()
}
}
