% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip.R
\name{peers.ip.collect}
\alias{peers.ip.collect}
\title{Collect connected peers' IP addresses}
\usage{
peers.ip.collect(
  csv.file = "xmr-peers-ip.csv",
  unrestricted.rpc.url = "http://127.0.0.1:18081",
  malicious.ips = NULL,
  top.subnet.mask = 24,
  n.top.subnets = 10,
  poll.time = 30
)
}
\arguments{
\item{csv.file}{The name of the CSV file to write to and read from. If it
already exists, data will be appended to it and the whole file will be
used to compute top subnet information.}

\item{unrestricted.rpc.url}{URL and port of the \code{monerod} unrestricted RPC.
Default is \verb{http://127.0.0.1:18081}}

\item{malicious.ips}{A character vector of IP addresses that are suspected
to be malicious.}

\item{top.subnet.mask}{Numeric value. The IP address subnet mask to print
summary information about.}

\item{n.top.subnets}{Number of subnets to print summary information about.}

\item{poll.time}{How often, in seconds, to collect data from the local
Monero node. Default is 30 seconds.}
}
\value{
NULL (invisible)
}
\description{
Collects IP addresses of peers that the local node has
established outbound connections to. The time and set of IP addresses are
saved to a CSV file. These IP addresses are checked against an optional
set of suspected malicious IP addresses. Information about the share of
outbound connections to suspected malicious IP addresses is printed.
IP addresses are grouped by subnet and information is printed to check for
possible "subnet saturation" by malicious entities. This function is an
infinite loop. \code{ctrl + c} to interrupt the function.
}
\examples{
\dontrun{
suspected.malicious.ips <- readLines(
  "https://raw.githubusercontent.com/Boog900/monero-ban-list/refs/heads/main/ban_list.txt")
peers.ip.collect(malicious.ips = suspected.malicious.ips)
}
}
