% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txpool.R
\name{txpool.collect}
\alias{txpool.collect}
\title{Collect txpool archive data}
\usage{
txpool.collect(
  db.file = "xmr-txpool-archive.db",
  unrestricted.rpc.url = "http://127.0.0.1:18081"
)
}
\arguments{
\item{db.file}{Name and path of database file created by \link{txpool.init}.}

\item{unrestricted.rpc.url}{URL and port of the \code{monerod} unrestricted RPC.
Default is \verb{http://127.0.0.1:18081}}
}
\value{
NULL (invisible)
}
\description{
Queries the local Monero node for its txpool once per second.
The time of arrival of each transaction is saved to a database file. The
database file must first be created by \link{txpool.init}. This function executes
an infinite loop. Input \code{ctrl + c} to interrupt the function.
}
\examples{
\dontrun{
txpool.collect()
}
}
\seealso{
\link{txpool.init}, which create the database file and \link{txpool.export},
which exports the database contents to a CSV file.
}
