% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txpool.R
\name{txpool.export}
\alias{txpool.export}
\title{Export txpool archive to CSV}
\usage{
txpool.export(
  db.file = "xmr-txpool-archive.db",
  csv.filepath = "",
  begin.date = "1970-01-01",
  end.date = "2035-01-01"
)
}
\arguments{
\item{db.file}{File name/path of the txpool archive database.}

\item{csv.filepath}{File path of CSV file that will be created. Leave
default to save in current working directory.}

\item{begin.date}{Optional argument to restrict data export date range.
Use "YYYY-MM-DD" format.}

\item{end.date}{Optional argument to restrict data export date range.
Use "YYYY-MM-DD" format.}
}
\value{
NULL (invisible)
}
\description{
Exports transaction and block arrival times from a database file
to a CSV file. Must use \link{txpool.init} and \link{txpool.collect} first.
\link{txpool.export} can be used while \link{txpool.collect} is still running in a
separate R session.
}
\examples{
\dontrun{
txpool.export()
}
}
\seealso{
\link{txpool.init}, which create the database file and \link{txpool.collect},
which collects txpool data and saves it to the database file.
}
