% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txpool.R
\name{txpool.init}
\alias{txpool.init}
\title{Initialize txpool archive database}
\usage{
txpool.init(db.file = "xmr-txpool-archive.db")
}
\arguments{
\item{db.file}{Name and path of database file to be created.}
}
\value{
NULL (invisible)
}
\description{
Initializes a database file to record the arrival time of
transactions to the local Monero node.
}
\examples{
\dontrun{
txpool.init()
}
}
\seealso{
\link{txpool.collect}, which collects txpool data and saves it to the
database file and \link{txpool.export}, which exports the database contents to a
CSV file.
}
